+++ rollup.config.js
@@ -7,5 +7,10 @@ var external = Object
 module.exports = {
   entry: 'src/index.js',
   plugins: [buble()],
-  external: external
+  external: external,
+  interop: false,
+  targets: [
+    { dest: 'dist/rollup-plugin-cleanup.js', format: 'cjs' },
+    { dest: 'dist/rollup-plugin-cleanup.es.js', format: 'es' }
+  ]
 }