+++ test/expected/es7.js
@@ -1,0 +1,4 @@
+async function foo () {
+  return 'foo'
+}
+foo().then( value => console.log( value ) )