+++ test/fixtures/es7.js
@@ -1,0 +1,9 @@
+/*
+  ES7
+*/
+
+async function foo () {
+  return 'foo'
+}
+
+foo().then( value => console.log( value ) )