+++ test/fixtures/long_comment.js
@@ -1,0 +1,8 @@
+/*
+  This is a long comment
+
+  Nostrud minim nisi aliqua non aute dolor ullamco anim eu consectetur. Adipisicing eu consequat laboris reprehenderit exercitation. Irure sunt sit reprehenderit proident tempor ea non nisi velit pariatur nostrud non laboris nostrud. Enim culpa proident ullamco est. Do sint id nisi commodo irure nulla.
+
+  Fugiat esse culpa voluptate laborum ea dolor ut qui aliquip cillum irure. Labore qui officia pariatur consectetur eiusmod. Ea officia sit laboris ut. Velit esse quis aliquip reprehenderit. In commodo occaecat voluptate pariatur cillum. Ad amet in velit duis labore mollit labore reprehenderit occaecat.
+*/
+console.log('Hi')